Added checks to avoid 404 error
Added a check to avoid a 404 error which happened if the user didn’t
automatically accept webcam access. The `src` attribute of the `video`
tag was set to `undefined`, which led to a 404 error on the URL
`http://domain.tld/path_with_scanner/undefined`. It is now just set to
an empty string.